+++ doc/changelog.md
@@ -11,6 +11,7 @@ TWBlue Changelog
     * Fixed an issue that was preventing TWBlue to create more than one user timeline during startup.
     * TWBlue will display properly new paragraphs in mastodon posts.
     * In the session manager, Mastodon sessions are now displayed including the instance to avoid confusion.
+    * TWBlue will now read default visibility preferences when posting new statuses, and display sensitive content. These preferences can be set on the mastodon instance, in the account's preferences section. If you wish to change TWBlue's behavior and have it not read those preferences from your instance, but instead set the default public visibility and hide sensitive content, you can uncheck the Read preferences from instance checkbox in the account options.
 
 ## Changes in version 2022.12.13